refacror: delete search-container, add class Search
