+++ src/index.ts
@@ -4,7 +4,3 @@ import './main.scss';
 
 const view = new View();
 view.drawApp();
-
-// const app = new App();
-// app.start();
-console.log(22222222222225);
